Started work on using PlatformIO.
