+++ .gitignore
@@ -1,2 +1,5 @@
-
+.pio
+.vscode/.browse.c_cpp.db*
 .vscode/c_cpp_properties.json
+.vscode/launch.json
+.vscode/ipch
