--- display.h
@@ -1,5 +1,0 @@
-void oled_msgLn(char *str, unsigned int length);
-void oled_msg(const char *str, unsigned int length);
-void oled_printf(const char *fmt, ...);
-void oled_init(void);
-void oled_print(const char *msg);
