+++ include/serial.h
@@ -1,0 +1,2 @@
+extern void serial_read(void);
+extern void serial_printf(const char *fmt, ...);
