+++ platformio.ini
@@ -1,0 +1,22 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:nodemcuv2]
+platform = espressif8266
+board = nodemcuv2
+framework = arduino
+lib_extra_dirs = ~/Documents/Arduino/libraries
+lib_deps = 
+	adafruit/Adafruit SSD1306@^2.4.7
+	adafruit/Adafruit Unified Sensor@^1.1.4
+	adafruit/Adafruit BME280 Library@^2.2.1
+	bakercp/CRC32@^2.0.0
+	adafruit/Adafruit HTU31D Library@^1.1.0
+	knolleary/PubSubClient@^2.8
