+++ src/.vscode/arduino.json
@@ -1,0 +1,7 @@
+{
+    "sketch": "ESPMQTTSensor.ino",
+    "configuration": "xtal=80,vt=flash,exception=disabled,stacksmash=disabled,ssl=all,mmu=3232,non32xfer=fast,ResetMethod=nodemcu,CrystalFreq=26,FlashFreq=40,FlashMode=dout,eesz=1M64,led=2,sdk=nonosdk_190703,ip=lm2f,dbg=Disabled,lvl=None____,wipe=none,baud=115200",
+    "board": "esp8266:esp8266:generic",
+    "intelliSenseGen": "global"
+
+}
