+++ src/parse.cpp
@@ -1,3 +1,59 @@
+#include <Arduino.h>
+#include <stdint.h>
+#include <PubSubClient.h>
+#include <SPI.h>
+#include <Wire.h>
+#include <Adafruit_GFX.h>
+#include <Adafruit_SSD1306.h>
+#include <EEPROM.h>
+#include <CRC32.h>
+
+#include "errors.h"
+#include "structs.h"
+#include "config.h"
+// #include "parse.h"
+#include "sensors.h"
+#include "display.h"
+#include "mqtt.h"
+#include "serial.h"
+#include "eeprom.h"
+#include "tasks.h"
+#include "sensor_htu31.h"
+#include "sensor_bme280.h"
+
+struct msgCallbackList msgs[] = {
+  { "SYSTEM.INIT.DEVICE", parse_system_init_device },
+  { "SYSTEM.INIT.DEVICE.SENSOR", parse_system_init_device },
+  { "SYSTEM.INIT.DEVICE.CONTROL", parse_system_init_device },
+  { "DEVICE.CONFIG.WIFI", parse_device_config_wifi },
+  { "DEVICE.CONFIG.NAME", parse_device_config_name },
+  { "DEVICE.CONFIG.LOCATION", parse_device_config_location },
+  { "DEVICE.CONFIG.MQTT", parse_device_config_mqtt },
+  { "DEVICE.REBOOT", parse_device_reboot },
+  { "DEBUG.CONFIG.DUMP", parse_debug_config_dump },
+  { "DEBUG.CONFIG.RDUMP", parse_debug_rconfig_dump },
+  { "DEBUG.GET.SENSOR", parse_debug_get_sensor },
+  { "DEVICE.CONFIG.CLEAR", parse_device_config_clear },
+  { "DEVICE.MQTT.SUBSCRIBE", parse_device_mqtt_subscribe },
+  { "GET.SENSOR", parse_config_stub },
+  { "UPDATE.SENSOR", parse_config_stub },
+  { "GET.STATE", parse_config_stub },
+  { "SET.STATE", parse_config_stub },
+  { NULL, NULL },
+};
+
+void parse_system_init_device(char *, char *[], unsigned int);
+void parse_config_stub(char *, char *[], unsigned int);
+void parse_device_config_name(char *, char *[], unsigned int);
+void parse_device_config_location(char *, char *[], unsigned int);
+void parse_device_reboot(char *, char *[], unsigned int);
+void parse_device_config_wifi(char *topic, char *argv[], unsigned int argc);
+void parse_device_config_mqtt(char *topic, char *argv[], unsigned int argc);
+void parse_device_config_clear(char *topic, char *argv[], unsigned int argc);
+void parse_device_mqtt_subscribe(char *topic, char *argv[], unsigned int argc);
+void parse_debug_config_dump(char *topic, char *argv[], unsigned int argc);
+void parse_debug_rconfig_dump(char *topic, char *argv[], unsigned int argc);
+void parse_debug_get_sensor(char *topic, char *argv[], unsigned int argc);
 
 void parse_system_init_device(char *topic, char *argv[], unsigned int argc) {
   serial_printf("\nEntered parsing function for %s\n", argv[0]);
