+++ src/serial.cpp
@@ -1,3 +1,34 @@
+#include <Arduino.h>
+#include <PubSubClient.h>
+#include <SPI.h>
+#include <Wire.h>
+#include <Adafruit_GFX.h>
+#include <Adafruit_SSD1306.h>
+#include <EEPROM.h>
+#include <CRC32.h>
+
+#include <stdint.h>
+#include <stdio.h>
+
+#include "errors.h"
+#include "structs.h"
+#include "config.h"
+#include "parse.h"
+#include "sensors.h"
+#include "display.h"
+#include "mqtt.h"
+// #include "serial.h"
+#include "eeprom.h"
+#include "tasks.h"
+#include "sensor_htu31.h"
+#include "sensor_bme280.h"
+
+void serial_read(void);
+void serial_printf(const char *fmt, ...);
+
+char serial_command_buffer[256];
+uint8_t serial_command_buffer_ptr = 0;
+
 void serial_read(void) {
   while (Serial.available() > 0) {
     uint8_t c = Serial.read();
